changed int by long long, fixed wrong operand order

+++ ponomarev.viacheslav/S2/computationalFunctions.hpp
@@ -12,7 +12,7 @@ namespace ponomarev
   ExpressionElement getElem(std::string & expression);
   bool isOpenBracketUp(const Stack< ExpressionElement > & stack);
   bool shouldPushOpToStack(const Stack< ExpressionElement > & stack, const ExpressionElement & elem);
-  int calculate(std::string & expression);
+  long long calculate(std::string & expression);
 }
 
 #endif
